+++ plydata/__init__.py
@@ -3,3 +3,20 @@ from ._version import get_versions
 
 __version__ = get_versions()['version']
 del get_versions
+
+
+def _get_all_imports():
+    """
+    Return list of all the imports
+    """
+    # 1. No local variables
+    # 2.`from Module import Something`, puts Module in
+    #    the namespace. We do not want that
+    import types
+    lst = [name for name, obj in globals().items()
+           if not (name.startswith('_') or
+                   isinstance(obj, types.ModuleType))]
+    return lst
+
+
+__all__ = _get_all_imports()